requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.0.6
Flask-Session==0.8.0
Flask-WTF==1.2.0
requests==2.32.3